chore: move jsonwall to a separate package
