+++ jsonwall/suite_test.go
@@ -5,7 +5,7 @@ import (
 
 	. "gopkg.in/check.v1"
 
-	"github.com/canonical/chisel/internal/jsonwall"
+	"github.com/canonical/chisel/jsonwall"
 )
 
 func Test(t *testing.T) { TestingT(t) }
